Make package lifetime_annotations compile under macOS.
Specifically, this is with the Blaze options

  --config=darwin_x86_64 --macos_minimum_os=10.14

This is needed to unblock integrating package lifetime_analysis into clang_tidy
(see unknown commit and presubmit failures there).

PiperOrigin-RevId: 415477220

+++ lifetime_annotations/lifetime_symbol_table.cc
@@ -63,7 +63,7 @@ static std::string NameFromIndex(int index) {
   std::string name;
   name.reserve(num_chars);
   for (int i = 0; i < num_chars; ++i) {
-    name.insert(0, 1, 'a' + index % 26);
+    name.insert(0, static_cast<size_t>(1), 'a' + index % 26);
     index /= 26;
   }
   return name;
